pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentforge"
version = "0.1.0"
description = "AI agent harness with pluggable memory strategies and built-in evaluation"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "MD Rabbi", email = "mrifat205@gmail.com"}]
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "tiktoken>=0.7.0",
    "rich>=13.0",
    "click>=8.0",
    "docker>=7.0",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.5",
    "pre-commit>=3.0",
]
rag = [
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
]

[project.scripts]
agentforge = "agentforge.cli:cli"

[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = ["agentforge/tests", "agentforge/agentforge"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501", "UP042"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
